Implementar autorización y filtrado de proyectos por manager

+++ src/controllers/ProjectController.ts
@@ -5,16 +5,24 @@ export class ProjectController {
     static createProject = async (req: Request, res: Response) => {
         const project = new Project(req.body)
 
+        // Asigna un manager
+        project.manager = req.user._id
+
         try {
             await project.save()
-            res.send('Proyecto creado correctamente')
+            res.send('Proyecto Creado Correctamente')
         } catch (error) {
             console.log(error)
         }
     }
     static getAllProjects = async (req: Request, res: Response) => {
         try {
-            const projects = await Project.find({})
+            const projects = await Project.find({
+                // Lógica para que solo se muestren los proyectos del usuario autenticado
+                $or: [
+                    { manager: {$in: req.user._id} }
+                ]
+            })
             res.json(projects)
         } catch (error) {
             res.status(500).json({ error: 'Hubo un error' })
@@ -28,6 +36,10 @@ export class ProjectController {
             if (!project) {
                 const error = new Error('Proyecto no encontrado')
                 return res.status(404).json({ error: error.message })
+            }
+            if(project.manager.toString() !== req.user._id.toString()) {
+                const error = new Error('No tienes permiso para ver este proyecto')
+                return res.status(403).json({ error: error.message })
             }
             res.json(project)
         } catch (error) {
@@ -43,6 +55,11 @@ export class ProjectController {
                 const error = new Error('Proyecto no encontrado')
                 return res.status(404).json({ error: error.message })
             }
+            if(project.manager.toString() !== req.user._id.toString()) {
+                const error = new Error('Solo el manager puede editar el proyecto')
+                return res.status(403).json({ error: error.message })
+            }
+
             project.clientName = req.body.clientName
             project.description = req.body.description
             project.projectName = req.body.projectName
@@ -60,6 +77,10 @@ export class ProjectController {
             if (!project) {
                 const error = new Error('Proyecto no encontrado')
                 return res.status(404).json({ error: error.message })
+            }
+            if(project.manager.toString() !== req.user._id.toString()) {
+                const error = new Error('Solo el manager puede eliminar el proyecto')
+                return res.status(403).json({ error: error.message })
             }
 
             await project.deleteOne()
